+++ src/main/java/gdsc/konkuk/platformcore/controller/member/MemberController.java
@@ -6,7 +6,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import gdsc.konkuk.platformcore.application.member.exceptions.MemberRegisterRequest;
+import gdsc.konkuk.platformcore.application.member.MemberRegisterRequest;
 import gdsc.konkuk.platformcore.application.member.MemberService;
 import gdsc.konkuk.platformcore.global.responses.SuccessResponse;
 import lombok.RequiredArgsConstructor;
